// frontend/src/services/api.d.ts
export interface components {
    schemas: {
// ...
        LoginResponse: {
            token?: string;
            /** Format: int64 */
            userId?: number;
            username?: string;
            realName?: string;
            phone?: string;
            email?: string;
            avatar?: string;
            /** Format: int32 */
            role?: number;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}